change(autostore): move container
